Add test module, move tests
